Finished variable edit costs

+++ src/hamming.rs
@@ -98,7 +98,7 @@ pub fn hamming_search_naive_k(needle: &[u8], haystack: &[u8], k: u32, best: bool
             }
         }
 
-        res.push(Match{start: i, end: i + needle_len, k: final_res as i32});
+        res.push(Match{start: i, end: i + needle_len, k: final_res});
 
         if best {
             curr_k = final_res;
@@ -106,7 +106,7 @@ pub fn hamming_search_naive_k(needle: &[u8], haystack: &[u8], k: u32, best: bool
     }
 
     if best {
-        res.retain(|m| m.k == curr_k as i32);
+        res.retain(|m| m.k == curr_k);
     }
 
     res
@@ -433,7 +433,7 @@ unsafe fn hamming_search_simd_core<T: Jewel>(needle: &[u8], haystack: &[u8], k: 
         let final_res = T::vector_count_mismatches(&needle_vector, haystack_ptr.offset(i as isize));
 
         if final_res <= curr_k {
-            res.push(Match{start: i, end: i + needle_len, k: final_res as i32});
+            res.push(Match{start: i, end: i + needle_len, k: final_res});
 
             if best {
                 curr_k = final_res;
@@ -452,7 +452,7 @@ unsafe fn hamming_search_simd_core<T: Jewel>(needle: &[u8], haystack: &[u8], k: 
             }
         }
 
-        res.push(Match{start: i, end: i + needle_len, k: final_res as i32});
+        res.push(Match{start: i, end: i + needle_len, k: final_res});
 
         if best {
             curr_k = final_res;
@@ -460,7 +460,7 @@ unsafe fn hamming_search_simd_core<T: Jewel>(needle: &[u8], haystack: &[u8], k: 
     }
 
     if best {
-        res.retain(|m| m.k == curr_k as i32);
+        res.retain(|m| m.k == curr_k);
     }
 
     res
